InitEventObjectStateFromTemplate

+++ include/global.fieldmap.h
@@ -59,8 +59,8 @@ struct EventObjectTemplate
     /*0x06*/ s16 y;
     /*0x08*/ u8 elevation;
     /*0x09*/ u8 movementType;
-    /*0x0A*/ u8 movementRangeX:4;
-             u8 movementRangeY:4;
+    /*0x0A*/ u16 movementRangeX:4;
+             u16 movementRangeY:4;
     /*0x0C*/ u16 trainerType;
     /*0x0E*/ u16 trainerRange_berryTreeId;
     /*0x10*/ const u8 *script;
